Major changes including functions.py and auto locker
A big update to add new features and overall changes.
- Added a new auto locker function which searches the deals page of OzBargain and the ProjectZeroThree website every 30 minutes for Unleaded 98 (at the moment) prices which are less than your wanted price, and if it is locks it in.
- Moved most of the functions from app.py to a new file called functions.py.
- Corrected the Device ID so it is 16 digits long.
