+++ .gitignore
@@ -4,3 +4,4 @@
 /__pycache__/*
 settings.py
 settings.py
+autolock.ini
